12,05

+++ DB/src/main/java/model/Faculty.java
@@ -13,7 +13,7 @@ public class Faculty implements Option {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "faculty_id")
-	private int id;
+	private long id;
 
 	private String faculty;
 
@@ -21,16 +21,16 @@ public class Faculty implements Option {
 		super();
 	}
 
-	public Faculty(int id) {
+	public Faculty(long id) {
 		super();
 		this.id = id;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
