+++ DB/src/main/java/model/FormOfEducation.java
@@ -13,10 +13,10 @@ public class FormOfEducation implements Option {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "foe_id")
-	private int id;
+	private long id;
 	private String foe;
 
-	public FormOfEducation(int id) {
+	public FormOfEducation(long id) {
 		super();
 		this.id = id;
 	}
@@ -25,11 +25,11 @@ public class FormOfEducation implements Option {
 		super();
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
